Do not use deprecated api: avoid processing points
The deprecated API loads and sorts the point definitions:
https://github.com/aantron/bisect_ppx/blob/master/src/runtime/common.ml#L130

We do not use the points definitions, just the counts, so avoid
processing the points.

Also the deprecated API grouped counts by file basename, which may not
be unique across the entire project.
The new API returns a proper path for each file.

This speeds up the processing of 3778 bisect files from ~3 minutes to under 10 seconds.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>